+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+
+- Add support for hiding columns in Auto/BaseTable via the [MUI Hidden Component](https://material-ui-next.com/api/hidden/).
+- Fix a header padding issue with AutoTable
+
 ## 0.2.0
 
 - Upgrade to perch-data v0.12.0 [ Breaking: Now uses a Provider ]